fix: code refactoring for stats type
